ai_app/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.6"
dependencies = [
    "fastapi[standard]==0.115.5",
    "langchain==0.3.9",
    "langchain-community==0.3.8",
    "langchain-groq==0.2.1",
    "langchain-openai==0.2.10",
    "langchain-redis==0.1.1",
    "python-dotenv==1.0.1",
    "pydantic==2.10.2",
    "redis==5.2.0",
    "langgraph==0.2.53",
    "uvicorn==0.32.1",
    "ipython>=8.30.0",
    "langchain-together>=0.2.0",
    "pandas>=2.2.3",
    "langchain-xai>=0.1.0",
    "langchain-huggingface>=0.1.2",
    "supabase>=2.10.0",
]
